requirements.txt
requests
pandas
sqlalchemy
psycopg2-binary
scikit-learn
joblib
fastapi
uvicorn[standard]
python-dotenv